+++ Services/IProductService.cs
@@ -5,6 +5,6 @@ namespace OnlineStore.Services
     public interface IProductService
     {
         List<Product> GetProducts();
-        Product? GetProductById(int id);
+        Product? GetProductById(long id);
     }
 }
