+++ Services/IReviewService.cs
@@ -4,7 +4,7 @@ namespace OnlineStore.Services
 {
     public interface IReviewService
     {
-        public List<Review> GetReviews(int productID, int? amount);
+        public List<Review> GetReviews(long productID, int? amount);
         public void AddReview(Review review);
     }
 }
